third_party/WebKit/LayoutTests/svg/dom/href-baseval-animval.html

<!doctype html>
<title>Test UseCounters for SVGURIReference.href.baseVal/animVal</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg width="0" height="0">
  <use xlink:href="#foo"/>
  <script>
    var target = document.querySelector("use");
    test(function() {
      var baseval = target.href.baseVal;
      assert_true(internals.isUseCounted(document, 758));
    }, "has SVGHrefBaseVal UseCounter"); // 758 == UseCounter::Feature::SVGHrefBaseVal
    test(function() {
      var animval = target.href.animVal;
      assert_true(internals.isUseCounted(document, 759));
    }, "has SVGHrefAnimVal UseCounter"); // 759 == UseCounter::Feature::SVGHrefAnimVal
  </script>
</svg>
